fix: expected int but found float

+++ server/portfolio/Models/streamresponse.fs
@@ -3,7 +3,7 @@
 open System.Collections.Generic
 
 type public streamresponse() =
-    member val public chunk: int[] = null with get, set
+    member val public chunk: float[] = null with get, set
     member val public totalchunks: int64 = 0L with get, set
     member val public bitdepth: int = 0 with get, set
     member val public samplerate: int = 0 with get, set
